+++ tests/test_checker/test_noqa.py
@@ -120,6 +120,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS344': 1,
     'WPS345': 1,
     'WPS346': 1,
+    'WPS347': 3,
 
     'WPS400': 0,
     'WPS401': 0,
@@ -168,6 +169,8 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS444': 1,
     'WPS445': 1,
     'WPS446': 1,
+    'WPS447': 1,
+    'WPS448': 1,
 
     'WPS500': 1,
     'WPS501': 1,
@@ -205,6 +208,9 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS610': 1,
     'WPS611': 1,
     'WPS612': 1,
+
+    'WPS701': 2,
+    'WPS702': 2,
 })
 
 
@@ -212,6 +218,7 @@ def _assert_errors_count_in_output(output, errors, all_violations):
     found_errors = Counter(
         (match.group(0) for match in ERROR_PATTERN.finditer(output)),
     )
+
     for violation in all_violations:
         key = 'WPS{0}'.format(str(violation.code).zfill(3))
         assert key in errors, 'Unlisted #noqa violation'