+++ tests/test_visitors/test_tokenize/test_conditions/test_implicit_elif.py
@@ -46,6 +46,37 @@ else:
         ...
 """
 
+# False positives, triggered by other constructs with `else` clause
+# (see https://github.com/wemake-services/wemake-python-styleguide/issues/835)
+
+for_else = """
+for a in b:
+   ...
+else:
+   if some:
+       ...
+"""
+
+try_except_else = """
+try:
+   ...
+except:
+   ...
+else:
+   if some:
+       ...
+"""
+
+embedded_else = """
+... if ... else ...
+"""
+
+technically_correct_token_stream = """
+    42
+        a
+    else
+"""
+
 
 @pytest.mark.parametrize('code', [
     elif_cases,
@@ -83,3 +114,24 @@ def test_implicit_elif_statements(
     visitor.run()
 
     assert_errors(visitor, [ImplicitElifViolation])
+
+
+@pytest.mark.parametrize('code', [
+    for_else,
+    try_except_else,
+    embedded_else,
+    technically_correct_token_stream,
+])
+def test_false_positives_are_ignored(
+    code,
+    assert_errors,
+    parse_tokens,
+    default_options,
+):
+    """Testing regular conditions."""
+    file_tokens = parse_tokens(code)
+
+    visitor = IfElseVisitor(default_options, file_tokens=file_tokens)
+    visitor.run()
+
+    assert_errors(visitor, [])