+++ wemake_python_styleguide/violations/best_practices.py
@@ -65,8 +65,9 @@ Summary
    UnhashableTypeInHashViolation
    WrongKeywordConditionViolation
    WrongNamedKeywordViolation
-   MisRefactoredAssignmentViolation
-
+   MisrefactoredAssignmentViolation
+   ImplicitPrimitiveViolation
+   ApproximateConstantViolation
 Best practices
 --------------
 
@@ -117,7 +118,6 @@ Best practices
 .. autoclass:: WrongKeywordConditionViolation
 .. autoclass:: WrongNamedKeywordViolation
 .. autoclass:: MisRefactoredAssignmentViolation
-
 """
 
 from typing_extensions import final
@@ -228,6 +228,11 @@ class OveruseOfNoqaCommentViolation(SimpleViolation):
     Solution:
         Refactor your code to match our style.
         Or use a config file to switch off some checks.
+
+    Configuration:
+        This rule is configurable with ``--max-noqa-comments``.
+        Default:
+        :str:`wemake_python_styleguide.options.defaults.MAX_NOQA_COMMENTS`
 
     .. versionadded:: 0.7.0
 
@@ -1749,6 +1754,7 @@ class WrongKeywordConditionViolation(ASTViolation):
     code = 444
 
 
+@final
 class WrongNamedKeywordViolation(ASTViolation):
     """
     Forbids to have wrong named keywords in starred dicts.
@@ -1777,36 +1783,109 @@ class WrongNamedKeywordViolation(ASTViolation):
 
 
 @final
-class MisRefactoredAssignmentViolation(ASTViolation):
+class ImplicitPrimitiveViolation(ASTViolation):
     """
-    Forbids to have statements that do nothing.
+    Forbids to use implicit primitives in a form of ``lambda`` functions.
 
     Reasoning:
-        Statements that just access the value or expressions
-        used as statements indicate that your code
-        contains deadlines. They just pollute your codebase and do nothing.
+        When you use ``lambda`` that returns a primitive value
+        and takes no arguments, it means that
+        you should use a primitive type instead.
 
     Solution:
-        Refactor your code in case it was a typo or error.
-        Or just delete this code.
+        Replace ``lambda`` with ``int``, ``float``,
+        ``list``, or any other primitive.
 
     Example::
 
         # Correct:
-        def some_function():
-            price = 8 + 2
-            return price
+        defaultdict(int)
 
         # Wrong:
-        def some_function():
-            8 + 2
-            print
+        defaultdict(lambda: 0)
 
-    .. versionadded:: 0.5.0
-    .. versionchanged:: 0.11.0
+    .. versionadded:: 0.13.0
+
+    """
+
+    code = 446
+    error_template = 'Found implicit primitive in a form of `lambda`'
+
+
+@final
+class ApproximateConstantViolation(ASTViolation):
+    """
+    Forbids to use approximate constants.
+
+    Reasoning:
+        Some constants are already defined.
+        No need to write them again, use existing values.
+        We just compare numbers as strings and raise this violation
+        when they start with the same chars.
+
+    Solution:
+        Use pre-defined constants.
+
+    Example::
+
+        # Correct:
+        from math import pi
+        random_number = 3.15
+        too_short = 3.1
+
+        # Wrong:
+        pi = 3.14
+
+    See
+    :py:data:`~wemake_python_styleguide.constants.MATH_APPROXIMATE_CONSTANTS`
+    for full list of math constants that we check for.
+
+    See also:
+        https://docs.python.org/3/library/math.html#constants
+
+    .. versionadded:: 0.13.0
+
+    """
+
+    code = 447
+    error_template = 'Found approximate constant: {0}'
+
+
+@final
+class MisrefactoredAssignmentViolation(ASTViolation):
+    """
+    Forbids to use approximate constants.
+
+    Reasoning:
+        Some constants are already defined.
+        No need to write them again, use existing values.
+        We just compare numbers as strings and raise this violation
+        when they start with the same chars.
+
+    Solution:
+        Use pre-defined constants.
+
+    Example::
+
+        # Correct:
+        from math import pi
+        random_number = 3.15
+        too_short = 3.1
+
+        # Wrong:
+        pi = 3.14
+
+    See
+    :py:data:`~wemake_python_styleguide.constants.MATH_APPROXIMATE_CONSTANTS`
+    for full list of math constants that we check for.
+
+    See also:
+        https://docs.python.org/3/library/math.html#constants
+
+    .. versionadded:: 0.13.0
 
     """
 
     error_template = 'Found self assignment  with refactored assignment'
-    code = 446
-    previous_codes = {445}
+    code = 448
+    previous_codes = {447}