+++ wemake_python_styleguide/violations/consistency.py
@@ -72,6 +72,7 @@ Summary
    ZeroDivisionViolation
    MeaninglessNumberOperationViolation
    OperationSignNegationViolation
+   VagueImportViolation
 
 Consistency checks
 ------------------
@@ -123,6 +124,7 @@ Consistency checks
 .. autoclass:: ZeroDivisionViolation
 .. autoclass:: MeaninglessNumberOperationViolation
 .. autoclass:: OperationSignNegationViolation
+.. autoclass:: VagueImportViolation
 
 """
 
@@ -1748,3 +1750,37 @@ class OperationSignNegationViolation(ASTViolation):
 
     error_template = 'Found wrong operation sign'
     code = 346
+
+
+@final
+class VagueImportViolation(ASTViolation):
+    """
+    Forbids imports that outside of the module may cause confusion.
+
+    Reasoning:
+        See ``datetime.*`` in code? You know that it's from datetime.
+        See ``BaseView`` in a Django project? You know where it is from.
+        See ``loads``? It can be anything: ``yaml``, ``toml``, ``json`` ...
+
+    Example::
+
+        # Correct:
+        import json
+
+        ...
+
+        json.loads(content)
+
+        # Wrong:
+        from json import loads
+
+        ...
+
+        loads(content)
+
+    .. versionadded:: 0.13.0
+
+    """
+
+    error_template = 'Found vague import that may cause confusion: {0}'
+    code = 347